+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Add a `debug` option ([#407](https://github.com/SassDoc/sassdoc/pull/407))
+
 ## 2.1.13
 
 * Default destination is relative to CWD ([#403](https://github.com/SassDoc/sassdoc/pull/403))